% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watershed.R
\name{morphoReconstruct}
\alias{morphoReconstruct}
\title{形态学重建}
\arguments{
\item{marker}{标记图像，通常是腐蚀或变换后的图像。}

\item{mask}{掩膜图像，定义了重建的极限。}

\item{kernel2}{结构元素的大小，用于定义膨胀操作的范围。}
}
\value{
返回形态学重建后的图像。
}
\description{
这个函数执行形态学重建，使用指定的标记和掩膜图像，以及给定的结构元素。
}
\examples{
# 假设有两个图像 marker 和 mask
# 运行形态学重建
result <- morphoReconstruct(marker, mask, 5)
}
